requirements.txt
numpy==1.13.1
tensorflow_gpu==1.3.0
scipy==0.18.1
sk_video==1.1.8
tables==3.3.0
pandas==0.19.2
Pillow==4.2.1
